+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## main(v0.9.0)
 
 - feat: | Worker | 支持多语言
+- feat: | Worker | `NO_LIMIT_SEND_ROLE` 配置支持多角色, 逗号分割
 
 ## v0.8.7